+++ src/parser/volume-parser.js
@@ -7,15 +7,19 @@
 
 import { Matrix4 } from "../../lib/three.es6.js";
 
+import { defaults } from "../utils.js";
 import Parser from "./parser.js";
 import Volume from "../surface/volume.js";
 
 
 function VolumeParser( streamer, params ){
 
-    Parser.call( this, streamer, params );
+    var p = params || {};
+
+    Parser.call( this, streamer, p );
 
     this.volume = new Volume( this.name, this.path );
+    this.voxelSize  = defaults( p.voxelSize, 1 );
 
 }