+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file, following t
 - volume slice representation including interpolation support
 - xplor/cns volume file parser
 - `colorVolume` parameter for surface representation of volume data
+- `voxelSize` parameter for volume file parser
+
 
 ### Changed
 - renamed `volume` parameter in molecular surface representation to `colorVolume`